+++ src/http2/structs.h
@@ -102,7 +102,7 @@ typedef struct HTTP2_STATES {
     h2_stream_t current_stream;
     uint32_t last_open_stream_id;
     uint8_t header_block_fragments[HTTP2_MAX_HBF_BUFFER];
-    uint8_t header_block_fragments_pointer;     //points to the next byte to write in
+    uint32_t header_block_fragments_pointer;     //points to the next byte to write in
     uint8_t waiting_for_end_headers_flag;       //bool
     uint8_t waiting_for_HEADERS_frame;
     uint8_t received_end_stream;
